third_party/blink/web_tests/external/wpt/css/mediaqueries/prefers-color-scheme-expected.txt

This is a testharness.js-based test.
PASS Should be parseable: '(prefers-color-scheme)'
FAIL Should be parseable: '(prefers-color-scheme: no-preference)' assert_true: expected true got false
FAIL Should be parseable: '(prefers-color-scheme: light)' assert_true: expected true got false
FAIL Should be parseable: '(prefers-color-scheme: dark)' assert_true: expected true got false
PASS Should not be parseable: '(prefers-color-scheme: 0)'
PASS Should not be parseable: '(prefers-color-scheme: none)'
PASS Should not be parseable: '(prefers-color-scheme: 10px)'
PASS Should not be parseable: '(prefers-color-scheme: dark 0)'
PASS Should not be parseable: '(prefers-color-scheme: dark light)'
PASS Should not be parseable: '(prefers-color-scheme: light/dark)'
FAIL Check that no-preference evaluates to false in the boolean context assert_equals: expected true but got false
Harness: the test ran to completion.
